Crate wiringpi

Modules

pin

thread

time

Structs

WiringPi

Functions

board_revision
This returns the board revision of the Raspberry Pi.

setup
This initialises the wiringPi system and assumes that the calling program is going to be using the wiringPi pin numbering scheme.

setup_gpio
This is identical to setup(), however it allows the calling programs to use the Broadcom GPIO pin numbers directly with no re-mapping.

setup_phys
This is identical to setup(), however it allows the calling programs to use the physical pin numbers on the P1 connector only.

setup_sys
This initialises the wiringPi system but uses the /sys/class/gpio interface rather than accessing the hardware directly.