Struct wiringpi::WiringPi

pub struct WiringPi<Pin>(_);

Methods

impl<P: Pin> WiringPi<P>

fn input_pin(&self, pin: u16) -> InputPin<P>

fn output_pin(&self, pin: u16) -> OutputPin<P>

fn millis(&self) -> u32

This returns a number representing the number if milliseconds since your program called one of the setup functions.

It returns an unsigned 32-bit number which wraps after 49 days.

fn micros(&self) -> u32

This returns a number representing the number if microseconds since your program called one of the setup functions.

It returns an unsigned 32-bit number which wraps after 71 minutes.

fn digital_write_byte(&self, byte: u8)

This writes the 8-bit byte supplied to the first 8 GPIO pins. It’s the fastest way to set all 8 bits at once to a particular value, although it still takes two write operations to the Pi’s GPIO hardware.

impl<P: Pwm + Pin> WiringPi<P>

fn pwm_pin(&self) -> PwmPin<P>

impl<P: GpioClock + Pin> WiringPi<P>

fn clock_pin(&self) -> ClockPin<P>