Struct wiringpi::pin::OutputPin

pub struct OutputPin<Pin>(_, _);

Methods

impl<P: Pin> OutputPin<P>

pub fn new(pin: c_int) -> OutputPin<P>

pub fn number(&self) -> c_int

pub fn digital_write(&self, value: Value)

Writes the value High or Low (1 or 0) to the given pin which must have been previously set as an output.

pub fn analog_write(&self, value: u16)

This writes the given value to the supplied analog pin. You will need to register additional analog modules to enable this function for devices such as the Gertboard.

impl<P: Pin + RequiresRoot> OutputPin<P>

pub fn into_soft_pwm(self) -> SoftPwmPin<P>

impl<P: Pin + RequiresRoot> OutputPin<P>

pub fn into_input(self) -> InputPin<P>

impl<P: Pin + Pwm> OutputPin<P>

pub fn into_pwm(self) -> PwmPin<P>

impl<P: Pin + GpioClock> OutputPin<P>

pub fn into_clock(self) -> ClockPin<P>